# internal/manifest/api/gen/types.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.3 DO NOT EDIT.
package gen

import (
	"time"

	openapi_types "github.com/oapi-codegen/runtime/types"
)

// Manifest defines model for Manifest.
type Manifest struct {
	AuthorEmail   string             `json:"author_email"`
	AuthorName    string             `json:"author_name"`
	Category      string             `json:"category"`
	CreatedAt     time.Time          `json:"created_at"`
	Description   string             `json:"description"`
	Icon          string             `json:"icon"`
	Id            openapi_types.UUID `json:"id"`
	MetaCreatedAt time.Time          `json:"meta_created_at"`
	Tags          []string           `json:"tags"`
	Title         string             `json:"title"`
	Version       string             `json:"version"`
}

// ListManifestsParams defines parameters for ListManifests.
type ListManifestsParams struct {
	Limit  *int `form:"limit,omitempty" json:"limit,omitempty"`
	Offset *int `form:"offset,omitempty" json:"offset,omitempty"`
}

// GetManifestsBySearchParams defines parameters for GetManifestsBySearch.
type GetManifestsBySearchParams struct {
	Query          string  `form:"query" json:"query"`
	AcceptLanguage *string `json:"Accept-Language,omitempty"`
}